- Big Maps: fixed road placing in map editor
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@3512 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Editor/editscreen.cpp
@@ -4336,7 +4336,7 @@ UINT32	EditScreenHandle( void )
 
 	//Calculate general mouse information
 	GetMouseXY( &sGridX, &sGridY );
-	iMapIndex = sGridY * WORLD_COLS + sGridX;
+	iMapIndex = (INT32)sGridY * WORLD_COLS + sGridX;
 
 	DetermineUndoState();
